Document text selection wrappers
